refactor(auto_search): extract aliases.rs (normalize/collect/match)
Third slice. Extracted the title-normalization + alias-collection +
target-matching cluster:

  pub fn normalize_title
  pub fn token_set / token_overlap_ratio
  pub fn collect_aliases / collect_sibling_aliases / collect_extended_aliases
  pub fn dedupe_strings
  pub fn matches_target
  pub struct SiblingRejectPrecompute (+ build)
  pub(super) fn sibling_match_rejects
  fn split_title_segments
  + 12 #[test] fns (split_segments_*, matches_target_*)

Visibility bumps in mod.rs to enable the sibling-to-sibling calls:
  episode_match              → pub(super) (called by aliases::matches_target)
  is_pack_candidate_relation → pub(super) (called by aliases::collect_sibling_aliases)
Both were already file-private; the bump is from `fn` to `pub(super) fn`.

mod.rs now has:
    pub use aliases::{
        collect_aliases, collect_extended_aliases, collect_sibling_aliases,
        dedupe_strings, matches_target, normalize_title, token_overlap_ratio, token_set,
    };
    use aliases::{SiblingRejectPrecompute, sibling_match_rejects};

External callers (main.rs, rss.rs) resolve unchanged via the `pub use`.

mod.rs: 6006 → 5316 lines. aliases.rs: 701 lines.
Tests still passing: 601.
